refactor: 클럽 리스트 재활용 가능하도록 수정
