+++ android/app/src/main/java/com/happy/friendogly/presentation/ui/club/add/ClubAddEvent.kt
@@ -1,6 +1,6 @@
 package com.happy.friendogly.presentation.ui.club.add
 
-import com.happy.friendogly.presentation.ui.club.model.clubfilter.ClubFilter
+import com.happy.friendogly.presentation.ui.club.common.model.clubfilter.ClubFilter
 
 sealed interface ClubAddEvent {
     sealed interface Navigation : ClubAddEvent {
