+++ android/app/src/main/java/com/happy/friendogly/presentation/ui/club/add/ClubAddViewModel.kt
@@ -20,8 +20,8 @@ import com.happy.friendogly.presentation.ui.club.filter.ClubFilterItemActionHand
 import com.happy.friendogly.presentation.ui.club.mapper.toDomain
 import com.happy.friendogly.presentation.ui.club.mapper.toGenders
 import com.happy.friendogly.presentation.ui.club.mapper.toSizeTypes
-import com.happy.friendogly.presentation.ui.club.model.ClubFilterSelector
-import com.happy.friendogly.presentation.ui.club.model.clubfilter.ClubFilter
+import com.happy.friendogly.presentation.ui.club.common.model.ClubFilterSelector
+import com.happy.friendogly.presentation.ui.club.common.model.clubfilter.ClubFilter
 import com.happy.friendogly.presentation.utils.addSourceList
 import kotlinx.coroutines.launch
 import okhttp3.MultipartBody
