+++ android/app/src/main/java/com/happy/friendogly/presentation/ui/club/common/adapter/club/ClubViewHolder.kt
@@ -1,7 +1,8 @@
-package com.happy.friendogly.presentation.ui.club.list.adapter.club
+package com.happy.friendogly.presentation.ui.club.common.adapter.club
 
 import androidx.recyclerview.widget.RecyclerView
 import com.happy.friendogly.databinding.ItemClubBinding
+import com.happy.friendogly.presentation.ui.club.common.ClubItemActionHandler
 import com.happy.friendogly.presentation.ui.club.list.ClubListActionHandler
 import com.happy.friendogly.presentation.ui.club.list.ClubListUiModel
 import com.happy.friendogly.presentation.ui.club.list.adapter.filter.FilterAdapter
@@ -20,7 +21,7 @@ class ClubViewHolder(
 
     fun bind(
         clubListUiModel: ClubListUiModel,
-        actionHandler: ClubListActionHandler,
+        actionHandler: ClubItemActionHandler,
     ) {
         binding.club = clubListUiModel
         binding.actionHandler = actionHandler
