+++ android/app/src/main/java/com/happy/friendogly/presentation/ui/club/common/model/ClubPet.kt
@@ -1,0 +1,3 @@
+package com.happy.friendogly.presentation.ui.club.common.model
+
+data class ClubPet(val imageUrl: String?)
