+++ android/app/src/main/java/com/happy/friendogly/presentation/ui/club/detail/ClubDetailActivity.kt
@@ -16,7 +16,7 @@ import com.happy.friendogly.presentation.ui.club.detail.adapter.DetailProfileAda
 import com.happy.friendogly.presentation.ui.club.list.ClubListFragment
 import com.happy.friendogly.presentation.ui.club.list.adapter.filter.FilterAdapter
 import com.happy.friendogly.presentation.ui.club.menu.ClubMenuBottomSheet
-import com.happy.friendogly.presentation.ui.club.model.clubfilter.ClubFilter
+import com.happy.friendogly.presentation.ui.club.common.model.clubfilter.ClubFilter
 import com.happy.friendogly.presentation.ui.club.modify.ClubModifyActivity
 import com.happy.friendogly.presentation.ui.club.modify.ClubModifyUiModel
 import com.happy.friendogly.presentation.ui.club.select.PetSelectBottomSheet
