+++ android/app/src/main/java/com/happy/friendogly/presentation/ui/club/detail/ClubDetailUiModel.kt
@@ -2,7 +2,7 @@ package com.happy.friendogly.presentation.ui.club.detail
 
 import com.happy.friendogly.presentation.ui.club.detail.model.ClubDetailProfileUiModel
 import com.happy.friendogly.presentation.ui.club.detail.model.ClubDetailViewType
-import com.happy.friendogly.presentation.ui.club.model.clubfilter.ClubFilter
+import com.happy.friendogly.presentation.ui.club.common.model.clubfilter.ClubFilter
 import com.happy.friendogly.presentation.ui.club.modify.ClubModifyUiModel
 import kotlinx.datetime.LocalDateTime
