+++ android/app/src/main/java/com/happy/friendogly/presentation/ui/club/filter/bottom/ClubFilterBottomSheet.kt
@@ -11,7 +11,7 @@ import com.happy.friendogly.databinding.BottomSheetGenderFilterSelectorBinding
 import com.happy.friendogly.presentation.base.observeEvent
 import com.happy.friendogly.presentation.ui.club.filter.ClubFilterEvent
 import com.happy.friendogly.presentation.ui.club.filter.ClubFilterViewModel
-import com.happy.friendogly.presentation.ui.club.model.clubfilter.ClubFilter
+import com.happy.friendogly.presentation.ui.club.common.model.clubfilter.ClubFilter
 
 class ClubFilterBottomSheet(
     private val clubFilterType: ClubFilter,
