+++ android/app/src/main/java/com/happy/friendogly/presentation/ui/club/list/ClubListActionHandler.kt
@@ -1,11 +1,8 @@
 package com.happy.friendogly.presentation.ui.club.list
 
-import com.happy.friendogly.presentation.ui.club.model.clubfilter.ClubFilter
+import com.happy.friendogly.presentation.ui.club.common.model.clubfilter.ClubFilter
 
 interface ClubListActionHandler {
-    fun loadClub(clubId: Long)
-
-    fun addClub()
 
     fun selectParticipationFilter()
