+++ android/app/src/main/java/com/happy/friendogly/presentation/ui/club/list/ClubListFragment.kt
@@ -13,9 +13,10 @@ import com.happy.friendogly.databinding.FragmentClubListBinding
 import com.happy.friendogly.presentation.base.BaseFragment
 import com.happy.friendogly.presentation.base.observeEvent
 import com.happy.friendogly.presentation.ui.MainActivityActionHandler
+import com.happy.friendogly.presentation.ui.club.common.ClubItemActionHandler
 import com.happy.friendogly.presentation.ui.club.filter.bottom.ClubFilterBottomSheet
 import com.happy.friendogly.presentation.ui.club.filter.bottom.ParticipationFilterBottomSheet
-import com.happy.friendogly.presentation.ui.club.list.adapter.club.ClubListAdapter
+import com.happy.friendogly.presentation.ui.club.common.adapter.club.ClubListAdapter
 import com.happy.friendogly.presentation.ui.club.list.adapter.selectfilter.SelectFilterAdapter
 
 class ClubListFragment : BaseFragment<FragmentClubListBinding>(R.layout.fragment_club_list) {
@@ -32,7 +33,7 @@ class ClubListFragment : BaseFragment<FragmentClubListBinding>(R.layout.fragment
         SelectFilterAdapter(viewModel as ClubListActionHandler)
     }
     private val clubAdapter: ClubListAdapter by lazy {
-        ClubListAdapter(viewModel as ClubListActionHandler)
+        ClubListAdapter(viewModel as ClubItemActionHandler)
     }
 
     override fun initViewCreated() {
