+++ android/app/src/main/java/com/happy/friendogly/presentation/ui/club/list/adapter/filter/FilterViewHolder.kt
@@ -2,7 +2,7 @@ package com.happy.friendogly.presentation.ui.club.list.adapter.filter
 
 import androidx.recyclerview.widget.RecyclerView
 import com.happy.friendogly.databinding.ItemFilterBinding
-import com.happy.friendogly.presentation.ui.club.model.clubfilter.ClubFilter
+import com.happy.friendogly.presentation.ui.club.common.model.clubfilter.ClubFilter
 
 class FilterViewHolder(
     private val binding: ItemFilterBinding,
