+++ android/app/src/main/java/com/happy/friendogly/presentation/ui/club/list/adapter/selectfilter/SelectFilterAdapter.kt
@@ -6,7 +6,7 @@ import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import com.happy.friendogly.databinding.ItemSelectedFilterMenuBinding
 import com.happy.friendogly.presentation.ui.club.list.ClubListActionHandler
-import com.happy.friendogly.presentation.ui.club.model.clubfilter.ClubFilter
+import com.happy.friendogly.presentation.ui.club.common.model.clubfilter.ClubFilter
 
 class SelectFilterAdapter(
     private val actionHandler: ClubListActionHandler,
