+++ android/app/src/main/java/com/happy/friendogly/presentation/ui/club/list/adapter/selectfilter/SelectFilterViewHolder.kt
@@ -3,7 +3,7 @@ package com.happy.friendogly.presentation.ui.club.list.adapter.selectfilter
 import androidx.recyclerview.widget.RecyclerView
 import com.happy.friendogly.databinding.ItemSelectedFilterMenuBinding
 import com.happy.friendogly.presentation.ui.club.list.ClubListActionHandler
-import com.happy.friendogly.presentation.ui.club.model.clubfilter.ClubFilter
+import com.happy.friendogly.presentation.ui.club.common.model.clubfilter.ClubFilter
 
 class SelectFilterViewHolder(
     private val binding: ItemSelectedFilterMenuBinding,
