--- android/app/src/main/java/com/happy/friendogly/presentation/ui/club/model/ClubPet.kt
@@ -1,3 +1,0 @@
-package com.happy.friendogly.presentation.ui.club.model
-
-data class ClubPet(val imageUrl: String?)
