+++ android/app/src/main/java/com/happy/friendogly/presentation/ui/club/select/PetSelectBottomSheet.kt
@@ -18,7 +18,7 @@ import com.happy.friendogly.R
 import com.happy.friendogly.application.di.AppModule
 import com.happy.friendogly.databinding.BottomSheetDogSelectorBinding
 import com.happy.friendogly.presentation.base.observeEvent
-import com.happy.friendogly.presentation.ui.club.model.clubfilter.ClubFilter
+import com.happy.friendogly.presentation.ui.club.common.model.clubfilter.ClubFilter
 import com.happy.friendogly.presentation.ui.club.select.adapter.PetSelectAdapter
 import kotlin.math.abs
