+++ android/app/src/main/res/layout/bottom_sheet_filter_size_selector.xml
@@ -4,7 +4,7 @@
 
     <data>
 
-        <import type="com.happy.friendogly.presentation.ui.club.model.clubfilter.ClubFilter" />
+        <import type="com.happy.friendogly.presentation.ui.club.common.model.clubfilter.ClubFilter" />
 
         <variable
             name="vm"
