+++ android/app/src/main/res/layout/bottom_sheet_participation_filter_selector.xml
@@ -5,7 +5,7 @@
 
     <data>
 
-        <import type="com.happy.friendogly.presentation.ui.club.model.clubfilter.ParticipationFilter" />
+        <import type="com.happy.friendogly.presentation.ui.club.common.model.clubfilter.ParticipationFilter" />
 
         <variable
             name="vm"
