+++ android/app/src/main/res/layout/fragment_club_list.xml
@@ -5,7 +5,7 @@
 
     <data>
 
-        <import type="com.happy.friendogly.presentation.ui.club.model.clubfilter.ClubFilter" />
+        <import type="com.happy.friendogly.presentation.ui.club.common.model.clubfilter.ClubFilter" />
 
         <variable
             name="vm"
