+++ android/app/src/main/res/layout/include_club_list_content.xml
@@ -4,13 +4,6 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools">
 
-    <data>
-
-        <variable
-            name="vm"
-            type="com.happy.friendogly.presentation.ui.club.list.ClubListViewModel" />
-    </data>
-
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rcv_club_list_club"
         android:layout_width="match_parent"
