+++ android/app/src/main/res/layout/item_selected_filter_menu.xml
@@ -11,7 +11,7 @@
 
         <variable
             name="clubFilter"
-            type="com.happy.friendogly.presentation.ui.club.model.clubfilter.ClubFilter" />
+            type="com.happy.friendogly.presentation.ui.club.common.model.clubfilter.ClubFilter" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
